+++ history.md
@@ -50,6 +50,7 @@ node starsky-tools/build-tools/app-version-update.js
 - [x]   (Changed) _Back-end_ "/api/thumbnail/{f}" status 409 is changed to status 210
 - [x]   (Added) _Front-end_  Zoom in detailView _issue #242_
 - [x]   (Fixed) _Back-end_  Delete large number of files gives exception _issue #281_
+- [x]   (Added) _Front-end_  Keyboard accelerator Command / Ctrl A _issue #247_
 
 # version 0.4.4 - 2021-01-10
 - [x]   (Security) _App_ npm audit fix node_modules/ini 1.3.8